# backend/openapi.yaml
openapi: 3.0.0
info:
  title: BrainFood API
  description: |
    REST API für die BrainFood Vokabel-App mit FSRS-5 Spaced Repetition Algorithmus.
    
    Diese API ermöglicht die Verwaltung von Boxen, Karteikarten und die intelligente
    Wiederholungsplanung basierend auf dem FSRS-5 Algorithmus.
  version: 1.0.0
  contact:
    name: BrainFood API Support
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://your-domain.com/api
    description: Production server
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/register
  schemas:
...
    Card:
      type: object
      properties:
        id:
          type: string
        boxId:
          type: string
        front:
          type: string
          description: Question or word to learn
        back:
          type: string
          description: Answer or translation
        tags:
          type: string
          nullable: true
          description: Comma-separated tags
        stability:
          type: number
          description: FSRS-5 stability value
        difficulty:
          type: number
          description: FSRS-5 difficulty value (0-1)
        reps:
          type: integer
          description: Number of successful repetitions
        lapses:
          type: integer
          description: Number of failures
        lastReviewAt:
          type: string
          format: date-time
          nullable: true
        due:
          type: string
          format: date-time
          description: Next due date for review
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ReviewRating:
      type: string
      enum: [again, hard, good, easy]
      description: User rating for card review
...
    ReviewRequest:
      type: object
      required:
        - rating
      properties:
        rating:
          $ref: '#/components/schemas/ReviewRating'
    ReviewResponse:
      type: object
      properties:
        card:
          $ref: '#/components/schemas/Card'
        nextDue:
          type: string
          format: date-time
        interval:
          type: integer
          description: Interval in days
    Error:
      type: object
      properties:
        error:
          type: string
paths:
...
  /boxes/{boxId}/reviews/next:
    get:
      tags:
        - Reviews
      summary: Get next due cards
      description: Returns the next card(s) that are due for review based on FSRS-5 scheduling
      parameters:
        - name: boxId
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 1
          description: Maximum number of cards to return
      responses:
        '200':
          description: List of due cards
          content:
            application/json:
              schema:
                type: object
                properties:
                  cards:
                    type: array
                    items:
                      $ref: '#/components/schemas/Card'
                  count:
                    type: integer
  /cards/{cardId}/review:
    post:
      tags:
        - Reviews
      summary: Submit a card review
      description: |
        Submits a review rating for a card and updates its FSRS-5 state.
        The algorithm calculates the next due date and updates stability/difficulty values.
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewRequest'
      responses:
        '200':
          description: Review processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
  /boxes/{boxId}/stats:
    get:
      tags:
        - Reviews
      summary: Get box statistics
      description: Returns statistics about cards and reviews in a box
      parameters:
        - name: boxId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Box statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  stats:
                    type: object
                    properties:
                      dueCount:
                        type: integer
                        description: Number of cards currently due
                      nextDue:
                        type: string
                        format: date-time
                        nullable: true
                        description: Date of next card due
                      totalCards:
                        type: integer
                      totalReviews:
                        type: integer
                      totalLapses:
                        type: integer
                      dailyReviews:
                        type: array
                        items:
                          type: object
                          properties:
                            date:
                              type: string
                              format: date-time
                            count:
                              type: integer
  /import/pdf:
    post:
      tags:
        - Import
      summary: Import cards from PDF
      description: |
        Uploads a PDF file and uses OpenAI to automatically generate flashcards.
        The PDF text is extracted, chunked, and sent to OpenAI for card generation.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - boxId
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF file to import
                boxId:
                  type: string
                  description: Target box ID
                sourceLanguage:
                  type: string
                  default: Deutsch
                  description: Source language for vocabulary
                targetLanguage:
                  type: string
                  default: Englisch
                  description: Target language for vocabulary
      responses:
        '201':
          description: Cards imported successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  cards:
                    type: array
                    items:
                      $ref: '#/components/schemas/Card'
        '400':
          description: Invalid input or file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /import/text:
    post:
      tags:
        - Import
      summary: Import cards from text
      description: |
        Takes raw text and uses OpenAI to automatically generate flashcards.
        The text is chunked and sent to OpenAI for card generation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - boxId
                - text
              properties:
                boxId:
                  type: string
                text:
                  type: string
                  description: Text content to generate cards from
                sourceLanguage:
                  type: string
                  default: Deutsch
                targetLanguage:
                  type: string
                  default: Englisch
      responses:
        '201':
          description: Cards imported successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  cards:
                    type: array
                    items:
                      $ref: '#/components/schemas/Card'